Updated the files to more appropriately use the node-pattern

+++ you_node2.js
@@ -1,5 +1,5 @@
 let sum = 0;
 for (let i = 2; i < process.argv.length; i++) {
-    sum += parseFloat(process.argv[i]);
+    sum += parseInt(process.argv[i]);
 }
 console.log(sum);
